use shorts instad of ints

+++ data2h.lua
@@ -10,12 +10,12 @@ local function outf(format,...) return out:write(format:format(...)) end
 outf([[
 struct %s_char {
   char c;
-  int x, y, w;
+  short x, y, w;
 };
 ]], arg[3])
 outf([[
 static const struct %s_data {
-  int h, char_count;
+  short h, char_count;
   struct %s_char chars[%i];
 } %s_data = { %i, %i,
 ]], arg[3], arg[3], #font.chars, arg[3], font.h, #font.chars)
